+++ internal/interp/console.go
@@ -5,13 +5,20 @@ import (
 	"github.com/ncode/portugol-go/internal/diag"
 )
 
-func (i *Interpreter) configureConsole(settings []ast.ConsoleStmt) error {
+func (i *Interpreter) configure(settings []ast.Stmt) error {
 	for _, setting := range settings {
-		if err := i.charge(setting.At); err != nil {
+		if err := i.charge(setting.Start()); err != nil {
 			return err
 		}
-		if err := i.options.Host.UseConsole(); err != nil {
-			return diag.Diagnostic{Code: diag.RHost, Pos: setting.At, Message: "cannot configure console display", Cause: err}
+		switch s := setting.(type) {
+		case *ast.ConsoleStmt:
+			if err := i.options.Host.UseConsole(); err != nil {
+				return diag.Diagnostic{Code: diag.RHost, Pos: s.At, Message: "cannot configure console display", Cause: err}
+			}
+		case *ast.FileInputStmt:
+			if err := i.configureFile(s); err != nil {
+				return err
+			}
 		}
 	}
 	return nil